extract install.base module; avoid circular import issues
