+++ demo/install/resources/library_add_checkouts.sql
