+++ demo/install/resources/library_without_checkouts.sql
