+++ demo/management/commands/remove_arxiv_pairs_log.py
@@ -2,7 +2,7 @@ import os
 
 from django.core.management import BaseCommand
 
-from demo.arxiv_skeleton import get_arxiv_db_and_schema_log_path
+from demo.install.arxiv_skeleton import get_arxiv_db_and_schema_log_path
 
 
 class Command(BaseCommand):
