+++ demo/middleware.py
@@ -3,9 +3,11 @@ import logging
 
 from django.conf import settings
 
-from demo.arxiv_skeleton import append_db_and_arxiv_schema_to_log
-from demo.install import customize_settings, create_demo_database, ARXIV, load_custom_explorations
 from demo.db_namer import get_name
+from demo.install.arxiv_skeleton import append_db_and_arxiv_schema_to_log
+from demo.install.base import ARXIV, create_demo_database
+from demo.install.custom_settings import customize_settings
+from demo.install.explorations import load_custom_explorations
 from mathesar.database.base import create_mathesar_engine
 from mathesar.models.base import Database
 from mathesar.state import reset_reflection
